@@ -1,11 +1,9 @@
 source 'https://rubygems.org'
 
-ruby '2.3.0'
-gem 'rails', '4.2.5'
+ruby '2.4.1'
+gem 'rails', '4.2.9'
 
-gem 'awesome_print'
 gem 'coffee-rails'
-gem 'dalli'
 gem 'devise'
 gem 'draper'
 gem 'icalendar'
@@ -14,26 +12,24 @@ gem 'haml-rails'
 gem 'jbuilder'
 gem 'jquery-rails'
 gem 'kaminari'
-gem 'memcachier'
 gem 'morrisjs-rails'
 gem 'pg'
 gem 'queue_classic'
-gem 'newrelic_rpm'
 gem 'rack-cache'
 gem 'raphael-rails'
 gem 'sass-rails'
-gem 'thin'
 gem 'turbolinks'
 gem 'uglifier'
-gem 'unicorn'
 
 group :development do
+  gem 'awesome_print'
+  gem 'capistrano'
+  gem 'capistrano-bundler'
+  gem 'capistrano-passenger'
+  gem 'capistrano-rails'
+  gem 'capistrano-rbenv'
   gem 'spring'
   gem 'spring-commands-rspec'
-end
-
-group :production do
-  gem 'rails_12factor'
 end
 
 group :test do
