@@ -16,17 +16,7 @@ Ultimate::Application.configure do
   # Don't care if the mailer can't send
   config.action_mailer.raise_delivery_errors = false
   config.action_mailer.default_url_options = { host: 'ultimate.dev' }
-
-  # config.action_mailer.delivery_method = :sendmail
-  config.action_mailer.delivery_method = :smtp
-  config.action_mailer.smtp_settings = {
-    :address        => 'smtp.sendgrid.net',
-    :port           => '587',
-    :authentication => :plain,
-    :user_name      => ENV['SENDGRID_USERNAME'],
-    :password       => ENV['SENDGRID_PASSWORD'],
-    :domain         => 'heroku.com'
-  }
+  config.action_mailer.delivery_method = :sendmail
 
   # Print deprecation notices to the Rails logger
   config.active_support.deprecation = :log
